% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cds_scerevisiae}
\alias{cds_scerevisiae}
\title{Coding sequences (CDS) of S. cerevisiae}
\format{
A DNAStringSet object with CDS of S. cerevisiae.
}
\usage{
data(cds_scerevisiae)
}
\description{
Data were obtained from Ensembl Fungi, and only CDS of primary transcripts
were included.
}
\examples{
data(cds_scerevisiae)
}
\keyword{datasets}
