% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_classification.R
\name{classify_genes}
\alias{classify_genes}
\title{Classify genes into unique modes of duplication}
\usage{
classify_genes(gene_pairs_list = NULL)
}
\arguments{
\item{gene_pairs_list}{List of classified gene pairs as returned
by \code{classify_gene_pairs()}.}
}
\value{
A list of 2-column data frames with variables \strong{gene}
and \strong{type} representing gene ID and duplication type, respectively.
}
\description{
Classify genes into unique modes of duplication
}
\examples{
data(sce_diamond)
data(sce_annotation)
blast_list <- sce_diamond
annotation <- sce_annotation
gene_pairs_list <- classify_gene_pairs(blast_list, annotation)
class_genes <- classify_genes(gene_pairs_list)
}
