% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_duplicate_classification.R
\name{classify_ssd_pairs}
\alias{classify_ssd_pairs}
\title{Classify small-scale duplication-derived gene pairs into subcategories}
\usage{
classify_ssd_pairs(ssd_pairs = NULL, annotation = NULL, proximal_max = 10)
}
\arguments{
\item{ssd_pairs}{A 2-column data frame with SSD-derived gene pairs.
This data frame can be obtained by filtering the output of
\code{get_wgd_pairs()} to keep only rows where type == "SSD".}

\item{annotation}{A processed GRanges object as in each element of the list
returned by \code{syntenet::process_input()}.}

\item{proximal_max}{Numeric scalar with the maximum distance (in number
of genes) between two genes to consider them as proximal duplicates.
Default: 10.}
}
\value{
A 3-column data frame with the variables:
\describe{
\item{dup1}{Duplicated gene 1}
\item{dup2}{Duplicated gene 2}
\item{type}{Duplication type, which can be either
"TD" (tandem duplication), "PD" (proximal duplication), and
"DD" (dispersed duplication).}
}
}
\description{
SSD-derived gene pairs are classified into tandem, proximal, and dispersed
duplicates (TD, PD, and DD, respectively).
}
\examples{
data(sce_annotation)
annotation <- sce_annotation[[1]]
data(sce_anchors)
data(sce_duplicates)
# Get SSD-derived gene pairs
dups <- get_wgd_pairs(sce_anchors, sce_duplicates)
ssd_pairs <- dups[dups$type == "SSD", 1:2]

# Classify SSD-derived gene pairs
ssd_classes <- classify_ssd_pairs(ssd_pairs, annotation)
}
