% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diamond_inter}
\alias{diamond_inter}
\title{Interspecies DIAMOND output for yeast species}
\format{
A list of data frames (length 1) containing the output of a
DIAMOND search of S. cerevisiae against C. glabrata (outgroup).
}
\usage{
data(diamond_inter)
}
\description{
This list contains a similarity search of S. cerevisiae against
C. glabrata, and it was obtained with \code{run_diamond()}.
}
\examples{
data(diamond_inter)
}
\keyword{datasets}
