% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diamond_intra}
\alias{diamond_intra}
\title{List of data frames containing intraspecies DIAMOND output for S. cerevisiae}
\format{
A list of data frames (length 1) containing the whole paranome of
S. cerevisiae resulting from intragenome similarity searches.
}
\usage{
data(diamond_intra)
}
\description{
List obtained with \code{run_diamond()}.
}
\examples{
data(diamond_intra)
}
\keyword{datasets}
