% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ka_ks_analyses.R
\name{find_intersect_mixtures}
\alias{find_intersect_mixtures}
\title{Find line intersect between pairs of Gaussian mixtures}
\usage{
find_intersect_mixtures(ks, peaks)
}
\arguments{
\item{ks}{A numeric vector of Ks values.}

\item{peaks}{A list with mean, standard deviation, and amplitude of Ks}
}
\value{
A numeric scalar or vector with the x-axis coordinates of the
intersections.
}
\description{
This function finds x-axis coordinate of n-1 intersections between lines
of n Gaussian mixtures. Thus, it will find 1 intersection for Ks distros
with 2 peaks, 2 intersections for distros with 2 peaks, and so on.
}
\examples{
data(gma_dups_kaks)
ks <- gma_dups_kaks$Ks
ks <- ks[!is.na(ks)]

# Remove Ks values >2 for testing purposes
ks <- ks[ks <= 2]

# Create list of peaks
peaks <- find_ks_peaks(ks, npeaks = 3)

# Get intersects
inter <- find_intersect_mixtures(ks, peaks)
}
