% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ka_ks_analyses.R
\name{find_peak_number}
\alias{find_peak_number}
\title{Find optimal number of peaks in Ks distribution}
\usage{
find_peak_number(
  ks,
  bootstraps = 1000,
  max_components = 2,
  signif = 0.05,
  epsilon = 0.001
)
}
\arguments{
\item{ks}{A numeric vector of Ks values.}

\item{bootstraps}{Numeric scalar with the number of bootstrap realizations
of the likelihood ratio statistic. We recommend using a minimum of 1000
bootstraps. Default: 1000.}

\item{max_components}{Numeric scalar with the maximum number of components
to test. Default: 2.}

\item{signif}{Numeric scalar with the significance level for the
likelihood ratio test. Default: 0.05.}

\item{epsilon}{Numeric scalar with the convergence criterion for the
expectation-maximization algorithm.}
}
\value{
A numeric scalar with the optimal number of peaks.
}
\description{
Find optimal number of peaks in Ks distribution
}
\examples{
data(gma_dups_kaks)
ks <- gma_dups_kaks$Ks
ks <- ks[!is.na(ks)]

# Remove Ks values > 1 for testing purposes
ks <- ks[ks <= 1]

# Using only 5 bootstraps for testing purposes
bootstraps <- 5
npeaks <- find_peak_number(ks, bootstraps, max_components = 2)

}
