% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gma_dups_kaks}
\alias{gma_dups_kaks}
\title{Soybean duplicated gene pairs and their Ka, Ks, and Ka/Ks values}
\format{
A data frame with the following columns:
\describe{
\item{dup1}{Character, duplicated gene 1.}
\item{dup2}{Character, duplicated gene 2.}
\item{Ka}{Numeric, Ka values.}
\item{Ks}{Numeric, Ks values.}
\item{Ka_Ks}{Numeric, Ka/Ks values.}
\item{mode}{Mode of duplication.}
}
}
\usage{
data(gma_dups_kaks)
}
\description{
Soybean duplicated gene pairs and their Ka, Ks, and Ka/Ks values
}
\examples{
data(gma_dups_kaks)
}
\references{
Almeida-Silva, F., Moharana, K. C., Machado, F. B., &
Venancio, T. M. (2020). Exploring the complexity of soybean (Glycine max)
transcriptional regulation using global gene co-expression networks.
Planta, 252(6), 1-12.
}
\keyword{datasets}
