% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmax_ks}
\alias{gmax_ks}
\title{Duplicate pairs and Ks values for Glycine max}
\format{
A data frame with the following variables:
\describe{
\item{dup1}{Character, duplicated gene 1.}
\item{dup2}{Character, duplicated gene 2.}
\item{Ks}{Numeric, Ks values.}
}
}
\usage{
data(gmax_ks)
}
\description{
This data set was obtained with \code{classify_gene_pairs()} followed
by \code{pairs2kaks()}.
}
\examples{
data(gmax_ks)
}
\keyword{datasets}
