% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ka_ks_analyses.R
\name{plot_ks_peaks}
\alias{plot_ks_peaks}
\title{Plot histogram of Ks distribution with peaks}
\usage{
plot_ks_peaks(ks, peaks = NULL, binwidth = 0.05)
}
\arguments{
\item{ks}{A numeric vector of Ks values.}

\item{peaks}{A list with mean, standard deviation, and amplitude of Ks
peaks as generated by \code{find_ks_peaks}.}

\item{binwidth}{Numeric scalar with binwidth for the histogram.
Default: 0.05.}
}
\value{
A ggplot object with a histogram and lines for each Ks peak.
}
\description{
Plot histogram of Ks distribution with peaks
}
\examples{
data(gma_dups_kaks)
ks <- gma_dups_kaks$Ks
ks <- ks[!is.na(ks)]

# Remove Ks values >1 for testing purposes
ks <- ks[ks <= 1]

# Create list of peaks
peaks <- find_ks_peaks(ks, npeaks = 2)

# Plot
plot_ks_peaks(ks, peaks, binwidth = 0.05)
}
