% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_anchors}
\alias{sce_anchors}
\title{Anchor pairs in the S. cerevisiae genome}
\format{
A 2-column data frame with each gene of the anchor pair in each
column. Columns are named "Anchor1" and "Anchor2".
}
\usage{
data(sce_anchors)
}
\description{
Anchor pairs were obtained with \code{get_anchors_list()}.
}
\examples{
data(sce_anchors)
}
\keyword{datasets}
