% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_annotation}
\alias{sce_annotation}
\title{Processed genome annotation for Saccharomyces cerevisiae}
\format{
A CompressedGRangesList containing
the element \code{Sce}.
}
\usage{
data(sce_annotation)
}
\description{
Data obtained from Pico-PLAZA 3.0. Only annotation data for primary
transcripts were included. The data was processed with
syntenet::process_input().
}
\examples{
data(sce_annotation)
}
\references{
Van Bel, M., Silvestri, F., Weitz, E. M., Kreft, L., Botzki, A.,
Coppens, F., & Vandepoele, K. (2021). PLAZA 5.0: extending the scope
and power of comparative and functional genomics in plants.
Nucleic acids research.
}
\keyword{datasets}
