% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_diamond}
\alias{sce_diamond}
\title{List of data frames containing BLAST-like tabular output}
\format{
A list of data frames (length 1) containing the whole paranome of
S. cerevisiae resulting from intragenome similarity searches.
}
\usage{
data(sce_diamond)
}
\description{
Only the DIAMOND output for S. cerevisiae against itself was included.
Here, we used DIAMOND in sensitive mode.
}
\examples{
data(sce_diamond)
}
\keyword{datasets}
