% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_duplicates}
\alias{sce_duplicates}
\title{All duplicate gene pairs for S. cerevisiae}
\format{
A 2-column data frame with each gene of the duplicate pair in each
column. Columns are named "dup1" and "dup2", respectively.
}
\usage{
data(sce_duplicates)
}
\description{
Duplicate gene pairs were obtained from the output of DIAMOND in sensitive
mode with an E-value filter of 1e-10.
}
\examples{
data(sce_duplicates)
}
\keyword{datasets}
