% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scerevisiae_kaks}
\alias{scerevisiae_kaks}
\title{Duplicate pairs and Ka, Ks, and Ka/Ks values for S. cerevisiae}
\format{
A data frame with the following variables:
\describe{
\item{dup1}{Character, duplicated gene 1.}
\item{dup2}{Character, duplicated gene 2.}
\item{Ka}{Numeric, Ka values.}
\item{Ks}{Numeric, Ks values.}
\item{Ka_Ks}{Numeric, Ka/Ks values.}
\item{type}{Character, mode of duplication}
}
}
\usage{
data(scerevisiae_kaks)
}
\description{
This data set was obtained with \code{classify_gene_pairs()} followed
by \code{pairs2kaks()}.
}
\examples{
data(scerevisiae_kaks)
}
\keyword{datasets}
