% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_seq}
\alias{yeast_seq}
\title{Protein sequences of the yeast species S. cerevisiae and C. glabrata}
\format{
A list of AAStringSet objects with the elements
\strong{Scerevisiae} and \strong{Cglabrata}.
}
\usage{
data(yeast_seq)
}
\description{
Data obtained from Ensembl Fungi. Only translated sequences of primary
transcripts were included.
}
\examples{
data(yeast_seq)
}
\keyword{datasets}
