% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_classification.R
\name{classify_genes}
\alias{classify_genes}
\title{Classify genes into unique modes of duplication}
\usage{
classify_genes(gene_pairs_list = NULL)
}
\arguments{
\item{gene_pairs_list}{List of classified gene pairs as returned
by \code{classify_gene_pairs()}.}
}
\value{
A list of 2-column data frames with variables \strong{gene}
and \strong{type} representing gene ID and duplication type, respectively.
}
\description{
Classify genes into unique modes of duplication
}
\examples{
data(diamond_intra)
data(yeast_annot)
data(yeast_seq)

pdata <- syntenet::process_input(yeast_seq, yeast_annot)
annot <- pdata$annotation["Scerevisiae"]
duplicates <- classify_gene_pairs(diamond_intra, annot)
class_genes <- classify_genes(duplicates)
}
