% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ka_ks_analyses.R
\name{find_ks_peaks}
\alias{find_ks_peaks}
\title{Find peaks in a Ks distribution by fitting Gaussian Mixture Models (GMMs)}
\usage{
find_ks_peaks(ks, npeaks = 2, epsilon = 0.001)
}
\arguments{
\item{ks}{A numeric vector of Ks values.}

\item{npeaks}{Numeric scalar indicating the number of peaks in
the distribution. If you don't know how many peaks there are,
the optimal number of peaks can be estimated with \code{find_peak_number}.
Default: 2.}

\item{epsilon}{Numeric scalar with the convergence criterion for the
expectation-maximization algorithm.}
}
\value{
A list with the following elements:
\describe{
\item{mean}{Numeric with the estimated means.}
\item{sd}{Numeric with the estimated standard deviations.}
\item{lambda}{Numeric with the estimated mixture weights.}
}
}
\description{
Find peaks in a Ks distribution by fitting Gaussian Mixture Models (GMMs)
}
\examples{
data(gma_dups_kaks)
ks <- gma_dups_kaks$Ks
ks <- ks[!is.na(ks)]

# Remove Ks values >1 for testing purposes
ks <- ks[ks <= 1]

# Find 2 peaks in Ks distribution
peaks <- find_ks_peaks(ks, npeaks = 2)
}
