% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_duplicate_classification.R
\name{get_transposed}
\alias{get_transposed}
\title{Get transposed duplicate pairs}
\usage{
get_transposed(pairs, blast_inter, annotation)
}
\arguments{
\item{pairs}{A 2-column data frame with duplicated gene 1 and 2 in
columns 1 and 2, respectively.}

\item{blast_inter}{A list of data frames of length 1
containing BLAST tabular output for the comparison between target
species and outgroup. Names of list elements must match the names of
list elements in \code{annotation}. BLASTp, DIAMOND or simular programs must
be run on processed sequence data as returned by \code{process_input()}.}

\item{annotation}{A processed GRangesList or CompressedGRangesList object as
returned by \code{syntenet::process_input()}.}
}
\value{
A 3-column data frame with the following variables:
\describe{
\item{dup1}{Duplicated gene 1}
\item{dup2}{Duplicated gene 2}
\item{type}{Duplication type, which can be either
"TRD" (transposed duplication) or
"DD" (dispersed duplication).}
}
}
\description{
Get transposed duplicate pairs
}
\examples{
data(diamond_inter)
data(diamond_intra)
data(yeast_seq)
data(yeast_annot)
blast_inter <- diamond_inter

# Get processed annotation
annotation <- process_input(yeast_seq, yeast_annot)$annotation

# Get duplicated pairs
annot <- pdata$annotation["Scerevisiae"] 
pairs_all <- classify_gene_pairs(diamond_intra, annot)
pairs <- pairs_all$Scerevisiae[pairs_all$Scerevisiae$type == "DD", ]

trd <- get_transposed(pairs, blast_inter, annotation)
}
