% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_duplicate_classification.R
\name{get_wgd_pairs}
\alias{get_wgd_pairs}
\title{Get gene pairs derived from whole-genome and small-scale duplications}
\usage{
get_wgd_pairs(anchor_pairs = NULL, duplicate_pairs = NULL)
}
\arguments{
\item{anchor_pairs}{A 2-column data frame with anchor pairs in columns 1
and 2.}

\item{duplicate_pairs}{A 2-column data frame with all duplicate pairs. This
is equivalent to the first 2 columns of the tabular output of BLAST-like
programs.}
}
\value{
A 3-column data frame with the variables:
\describe{
\item{dup1}{Duplicated gene 1}
\item{dup2}{Duplicated gene 2}
\item{type}{Duplication type, which can be either
"WGD" (whole-genome duplication) or
"SSD" (small-scale duplication).}
}
}
\description{
Get gene pairs derived from whole-genome and small-scale duplications
}
\examples{
data(sce_anchors)
data(sce_duplicates)
dups <- get_wgd_pairs(sce_anchors, sce_duplicates)
}
