% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_annot}
\alias{yeast_annot}
\title{Genome annotation of the yeast species S. cerevisiae and C. glabrata}
\format{
A CompressedGRangesList containing
the elements \strong{Scerevisiae} and \strong{Cglabrata}.
}
\usage{
data(yeast_annot)
}
\description{
Data obtained from Ensembl Fungi. Only annotation data protein-coding
genes (with associated mRNA, exons, CDS, etc) are included.
}
\examples{
data(yeast_annot)
}
\keyword{datasets}
